Rename python package to daxai and update imports
