+++ daxai/models/common.py
